refactor: rename and relocate files part 3 - update LinkedInIcon & FilterIcon
