+++ src/components/Directory/Filter/index.tsx
@@ -1,6 +1,7 @@
 import { useRef, useEffect } from "react";
 import { Industry } from "../../../app/types";
 import "./checkbox.css";
+import FilterIcon from "@/components/common/icons/Filter";
 
 interface FilterProps {
   industries: Industry[];
@@ -65,23 +66,7 @@ export default function Filter({
         } `}
       >
         <div className="flex items-center">
-          {/* Filter Icon SVG */}
-          <svg
-            width="20"
-            height="16"
-            viewBox="0 0 40 32"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-            className="text-black"
-          >
-            <path
-              d="M2 2L38 5.11111M6 8.22222L34 11.3333M10 14.4444L30 17.5556M14 20.6667L26 23.7778M18 26.8889L22 30"
-              stroke="currentColor"
-              strokeWidth="3"
-              strokeLinecap="round"
-            />
-          </svg>
-
+          <FilterIcon></FilterIcon>
           <span className="ml-2">Filter by Industry</span>
         </div>
         <div
