+++ src/components/common/icons/Filter/index.tsx
@@ -1,0 +1,21 @@
+const FilterIcon = () => {
+  return (
+    <svg
+      width="20"
+      height="16"
+      viewBox="0 0 40 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className="text-black"
+    >
+      <path
+        d="M2 2L38 5.11111M6 8.22222L34 11.3333M10 14.4444L30 17.5556M14 20.6667L26 23.7778M18 26.8889L22 30"
+        stroke="currentColor"
+        strokeWidth="3"
+        strokeLinecap="round"
+      />
+    </svg>
+  );
+};
+
+export default FilterIcon;
